Use alternate API for enabling SSE
This represents the SSE settings as an option.  This means all settings
are set up initially, instead of after the streamstore is created. While
it's not a problem for the sse setting, it might be for other things we
might want to configure soon (e.g, the profile is hard coded to "dev"
right now) so this establishes a good flexible way to do all options.