Standardise Seek API
Seek was avoiding the slightly unwields, C inspired Seek API used in the
standard library, in favour of a simpler one.  However, this makes straw
types harder to use and less of a "drop in" replacement for the usual
types in the Go standard library.

This adopts io.Seeker for seeking, although only io.SeekStart is
supported currently.